refactor: :recycle: rename file to exclude_non_sprout_resource_errors
