+++ tests/core/sprout_checks/test_exclude_non_sprout_resource_errors.py
@@ -1,5 +1,5 @@
 from seedcase_sprout.core.checks.check_error import CheckError
-from seedcase_sprout.core.sprout_checks.get_sprout_specific_resource_errors import (
+from seedcase_sprout.core.sprout_checks.exclude_non_sprout_resource_errors import (
     exclude_non_sprout_resource_errors,
 )
